Complete old PR suggestions
Read the PR from before this date for more detailed information.
Not listed on the PR:
Removed the momentum conservation feature of the shooter flywheel, which would keep running the flywheel when not actively shooting at a lower velocity to decrease ramp-up times.  This was never used and is a maintenance issue.
Removed two Prefs, SHOOTER_FLYWHEEL_SETPOINT and SHOOTER_FLYWHEEL_IDLE that were rendered useless by other changes documented on the PR.
Added a widget on the Match tab that allows the user to choose whether or not to enable PID on the shooter flywheel.