Complete PR Suggestions
See PR suggestions from the night of 5/7 for changes.
Changes not listed on PR:
Set the move_curve and turn_curve of CurvesConditioner to 2 to fix a failed unit test.
Change the arrangement of shootFromLine and shootFromWall to produce the desired action.
Fully nest control of the shooter hood inside ShooterSubsystem, so that RobotContainer doesn't have to interface with both Subsystems.