Remove turn request negation from DriveCommand
A negation was added in DriveCommand as a quick fix for an inverted turn
request. The proper way to fix this is to use a DriveConditioner.
This fix reverts the negation in DriveCommand and adds a new
DriveConditioner that inverts the turn request.

This fixes the broken DriveCommandTest unit test.