Add MoPrefsKey enum
MoPrefs are now accessed using the get() method, which takes a MoPrefsKey.
MoPrefsKeys are responsible for tracking the default value of their
associated preference, and are used as keys into the Preferences table.

This improves maintainability: adding a new preference only requires
adding a new entry to the MoPrefsKey enum instead of a whole new method
on MoPrefs.