Add paramterized constructor to CurvesConditioner
This patch adds a new constructor to the CurvesConditioner that allows the
move and turn curves to be defined when the class is constructed.

The old hard-coded values for the move and turn curves have been named
DEFAULT_MOVE_CURVE and DEFAULT_TURN_CURVE. These values are used by the
default constructor, to preserve backwards compatibility.

This change was necessary to fix the unit tests which were broken by the
last patch.