Implement tank drive
This implementation of tank drive differs from that in the tank-drive
branch in how it conditions the inputs.

This implementation calculates the arcade drive move_request and
turn_request from the tank drive inputs, then passes that into the
conditioner stack. This is desireable because it is much more useful to
condition lateral movement and rotation separately.